+++ bika/lims/browser/analysisrequest/publish.py
@@ -614,6 +614,9 @@ class AnalysisRequestPublishView(BrowserView):
                                        <Analysis (for as-1)>],
                              'ar2_id': [<Analysis (for as-1)>]
                             },
+                     'interims': {'ar1_id': [an_interims_report],
+                                  'ar2_id': [an_interims_report]
+                                 },
                     },
                 },
             {'service_2_title':
@@ -623,6 +626,9 @@ class AnalysisRequestPublishView(BrowserView):
                                        <Analysis (for as-2)>],
                              'ar2_id': [<Analysis (for as-2)>]
                             },
+                     'interims': {'ar1_id': [an_interims_report],
+                                  'ar2_id': [an_interims_report]
+                                 },
                     },
                 },
             ...
@@ -631,6 +637,8 @@ class AnalysisRequestPublishView(BrowserView):
         """
         analyses = {}
         for ar in ars:
+            an_interims_report = []
+            interims_per_ar = {}
             ans = [an.getObject() for an in ar.getAnalyses()]
             for an in ans:
                 cat = an.getCategoryTitle()
@@ -642,19 +650,24 @@ class AnalysisRequestPublishView(BrowserView):
                             # here - service fields are all inside!
                             'service': an,
                             'accredited': an.getAccredited(),
-                            'ars': {ar.id: an.getFormattedResult()}
+                            'ars': {ar.id: an.getFormattedResult()},
                         }
                     }
                 elif an_title not in analyses[cat]:
                     analyses[cat][an_title] = {
                         'service': an,
                         'accredited': an.getAccredited(),
-                        'ars': {ar.id: an.getFormattedResult()}
+                        'ars': {ar.id: an.getFormattedResult()},
                     }
                 else:
                     d = analyses[cat][an_title]
                     d['ars'][ar.id] = an.getFormattedResult()
                     analyses[cat][an_title] = d
+
+                interims = an.getInterimFields()
+                an_interims_report = filter(lambda interim: interim.get('report', False), interims)
+                interims_per_ar.update({ar.id: an_interims_report})
+                analyses[cat][an_title]['interims'] = interims_per_ar
         return analyses
 
     def _lab_address(self, lab):
@@ -691,15 +704,6 @@ class AnalysisRequestDigester:
     called for all subsequent digestion.  This allows the instance to cache
     data for objects that may be read multiple times for different ARs.
 
-    Passing overwrite=True when calling the instance will cause the
-    ar.Digest field to be overwritten with a new digestion.  This flag
-    is set True by default in the EndRequestHandler that is responsible for
-    automated re-building.
-
-    It should be run once when the AR is verified (or when a verified AR is
-    modified) to pre-digest the data so that AnalysisRequestPublishView will
-    run a little faster.
-
     Note: ProxyFields are not included in the reading of the schema.  If you
     want to access sample fields in the report template, you must refer
     directly to the correct field in the Sample data dictionary.
@@ -717,7 +721,7 @@ class AnalysisRequestDigester:
             'creators', 'effectiveDate', 'expirationDate', 'language', 'rights',
             'relatedItems', 'modification_date', 'immediatelyAddableTypes',
             'locallyAllowedTypes', 'nextPreviousEnabled', 'constrainTypesMode',
-            'RestrictedCategories', 'Digest',
+            'RestrictedCategories',
         ]
 
     def __call__(self, ar, overwrite=False):
@@ -725,38 +729,9 @@ class AnalysisRequestDigester:
         self.context = ar
         self.request = ar.REQUEST
 
-        # if AR was previously digested, use existing data (if exists)
-        verified = wasTransitionPerformed(ar, 'verify')
-        if not overwrite and verified:
-            # Prevent any error related with digest
-            data = ar.getDigest() if hasattr(ar, 'getDigest') else {}
-            if data:
-                # Check if the department managers have changed since
-                # verification:
-                saved_managers = data.get('managers', {})
-                saved_managers_ids = set(saved_managers.get('ids', []))
-                current_managers = self.context.getManagers()
-                current_managers_ids = set([man.getId() for man in
-                                            current_managers])
-                # The symmetric difference of two sets A and B is the set of
-                # elements which are in either of the sets A or B but not
-                # in both.
-                are_different = saved_managers_ids.symmetric_difference(
-                    current_managers_ids)
-                if len(are_different) == 0:
-                    # Seems that sometimes the 'obj' is wrong in the saved
-                    # data.
-                    data['obj'] = ar
-                    # Always set results interpretation
-                    self._set_results_interpretation(ar, data)
-                    return data
-
         logger.info("=========== creating new data for %s" % ar)
-
         # Set data to the AR schema field, and return it.
         data = self._ar_data(ar)
-        if hasattr(ar, 'setDigest'):
-            ar.setDigest(data)
         logger.info("=========== new data for %s created." % ar)
         return data
 
@@ -948,14 +923,13 @@ class AnalysisRequestDigester:
 
         # Sub-objects
         excludearuids.append(ar.UID())
-        puid = ar.getRawRetracted()
+        puid = ar.getRawInvalidated()
         if puid and puid not in excludearuids:
             data['parent_analysisrequest'] = self._ar_data(
-                ar.getRetracted(), excludearuids)
-        cuid = ar.getRawRetest()
-        if cuid and cuid not in excludearuids:
-            data['child_analysisrequest'] = self._ar_data(
-                ar.getRetest(), excludearuids)
+                ar.getInvalidated(), excludearuids)
+        retest = ar.getRetest()
+        if retest and api.get_uid(retest) not in excludearuids:
+            data['child_analysisrequest'] = self._ar_data(retest, excludearuids)
 
         wf = ar.portal_workflow
         allowed_states = ['verified', 'published']
@@ -968,6 +942,9 @@ class AnalysisRequestDigester:
         data['batch'] = self._batch_data(ar)
         data['specifications'] = self._specs_data(ar)
         data['analyses'] = self._analyses_data(ar, ['verified', 'published'])
+        data['hasinterimfields'] = len(
+            [an['interims'] for an in data['analyses'] if
+             len(an['interims']) > 0]) > 0
         data['qcanalyses'] = self._qcanalyses_data(ar,
                                                    ['verified', 'published'])
         data['points_of_capture'] = sorted(
@@ -1391,6 +1368,8 @@ class AnalysisRequestDigester:
         # is out of range. The second value (dismissed here) is a bool that
         # indicates if the result is out of shoulders
         andict['outofrange'] = is_out_of_range(analysis)[0]
+        interims = analysis.getInterimFields()
+        andict['interims'] = filter(lambda interim: interim.get('report', False), interims)
         return andict
 
     def _qcanalyses_data(self, ar, analysis_states=None):
@@ -1531,47 +1510,6 @@ class AnalysisRequestDigester:
         """Returns true if hidden analyses are visible
         """
         return self.request.form.get('hvisible', '0').lower() in ['true', '1']
-
-
-def ARModifiedHandler(instance, event):
-    """After any modification of an AR that has already been verified,
-    re-populate the ar.Digest.
-    """
-    if IAnalysisRequest.providedBy(instance):
-        if wasTransitionPerformed(instance, 'verify'):
-            request = instance.REQUEST
-            ars_to_digest = set(request.get('ars_to_digest', []))
-            ars_to_digest.add(instance)
-            request['ars_to_digest'] = ars_to_digest
-
-
-def AnalysisAfterTransitionHandler(instance, event):
-    """After a 'verify' transition on any analysis, we must set a flag in
-    the request so that the AR is digested before the request terminates.
-    We're doing it here so that the digestion happens only once at the end
-    of the request, regardless of how many children were transitioned.
-    """
-    if event.transition and event.transition.id == 'verify':
-        request = instance.REQUEST
-        ar = instance.aq_parent
-        ars_to_digest = set(request.get('ars_to_digest', []))
-        ars_to_digest.add(ar)
-        request['ars_to_digest'] = ars_to_digest
-
-
-def EndRequestHandler(event):
-    """At the end of the request, we check, to see if any pre-digestion is
-    required, for any ars or analyses that were processed during the request.
-    """
-    request = event.request
-    ars_to_digest = set(request.get('ars_to_digest', []))
-    digester = AnalysisRequestDigester()
-    if ars_to_digest:
-        for ar in ars_to_digest:
-            digester(ar, overwrite=True)
-    # If this commit() is not here, then the data does not appear to be
-    # saved.  IEndRequest happens outside the transaction?
-    transaction.commit()
 
 
 def get_client_address(context):