+++ bika/lims/upgrade/v01_02_009.py
@@ -6,9 +6,8 @@
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
 import time
-import transaction
-from Products.Archetypes.config import REFERENCE_CATALOG
 
+import transaction
 from bika.lims import api
 from bika.lims import logger
 from bika.lims.catalog.analysisrequest_catalog import \
@@ -16,6 +15,7 @@ from bika.lims.catalog.analysisrequest_catalog import \
 from bika.lims.config import PROJECTNAME as product
 from bika.lims.upgrade import upgradestep
 from bika.lims.upgrade.utils import UpgradeUtils
+from Products.Archetypes.config import REFERENCE_CATALOG
 from Products.DCWorkflow.Guard import Guard
 
 version = '1.2.9'  # Remember version number in metadata.xml and setup.py
@@ -41,8 +41,11 @@ def upgrade(tool):
     # Reindex object security for client contents (see #991)
     reindex_client_local_owner_permissions(portal)
 
-    # Rebind ARs that were generated because of the retraction of other ARs
-    rebind_retracted_ars(portal)
+    # Rename "retract_ar" transition to "invalidate"
+    rename_retract_ar_transition(portal)
+
+    # Rebind ARs that were generated because of the invalidation of other ARs
+    rebind_invalidated_ars(portal)
 
     # Add "Create partition" transition
     add_create_partition_transition(portal)
@@ -70,7 +73,7 @@ def reindex_client_local_owner_permissions(portal):
                 .format(end-start))
     transaction.commit()
 
-    
+
 def migrate_attachment_report_options(portal):
     """Migrate Attachments with the report option "a" (attach in report)
        to the option to "i" (ignore in report)
@@ -86,6 +89,81 @@ def migrate_attachment_report_options(portal):
             obj.setReportOption("i")
             obj.reindexObject()
             logger.info("Migrated Attachment %s" % obj.getTextTitle())
+
+
+def rename_retract_ar_transition(portal):
+    """Renames retract_ar transition to invalidate
+    """
+    logger.info("Renaming 'retract_ar' transition to 'invalidate'")
+    wf_tool = api.get_tool("portal_workflow")
+    workflow = wf_tool.getWorkflowById("bika_ar_workflow")
+
+    if "invalidate" not in workflow.transitions:
+        workflow.transitions.addTransition("invalidate")
+    transition = workflow.transitions.invalidate
+    transition.setProperties(
+        title="Invalidate",
+        new_state_id="invalid",
+        after_script_name="",
+        actbox_name="Invalidate",
+    )
+    guard = transition.guard or Guard()
+    guard_props = {"guard_permissions": "BIKA: Retract",
+                   "guard_roles": "",
+                   "guard_expr": "python:here.guard_cancelled_object()"}
+    guard.changeFromProperties(guard_props)
+    transition.guard = guard
+
+    for state in workflow.states.values():
+        if 'retract_ar' in state.transitions:
+            trans = filter(lambda id: id != 'retract_ar', state.transitions)
+            trans += ('invalidate', )
+            state.transitions = trans
+
+    if "retract_ar" in workflow.transitions:
+        workflow.transitions.deleteTransitions(["retract_ar"])
+
+
+def rebind_invalidated_ars(portal):
+    """Rebind the ARs automatically generated because of the retraction of their
+    parent to the new field 'Invalidated'. The field used until now
+    'ParentAnalysisRequest' will be used for partitioning
+    """
+    logger.info("Rebinding retracted/invalidated ARs")
+
+    # Walk through the Analysis Requests that were generated because of an
+    # invalidation, get the source AR and rebind the fields
+    relationship = "AnalysisRequestChildAnalysisRequest"
+    ref_catalog = api.get_tool(REFERENCE_CATALOG)
+    retests = ref_catalog(relationship=relationship)
+    total = len(retests)
+    to_remove = list()
+    num = 0
+    for num, relation in enumerate(retests, start=1):
+        relation = relation.getObject()
+        if not relation:
+            continue
+        retest = relation.getTargetObject()
+        invalidated = relation.getSourceObject()
+        retest.setInvalidated(invalidated)
+        # Set ParentAnalysisRequest field to None, cause we will use this field
+        # for storing Primary-Partitions relationship.
+        retest.setParentAnalysisRequest(None)
+
+        # Remove the relationship!
+        to_remove.append((relation.aq_parent, relation.id))
+
+        if num % 100 == 0:
+            logger.info("Rebinding invalidated ARs: {0}/{1}"
+                        .format(num, total))
+
+    # Remove relationships
+    for relation_to_remove in to_remove:
+        folder = relation_to_remove[0]
+        rel_id = relation_to_remove[1]
+        folder.manage_delObjects([rel_id])
+
+    logger.info("Rebound {} invalidated ARs".format(num))
 
 
 def add_create_partition_transition(portal):
@@ -132,38 +210,3 @@ def add_create_partition_transition(portal):
         workflow.updateRoleMappingsFor(api.get_object(ar))
         if num % 100 == 0:
             logger.info("Updating role mappings: {0}/{1}".format(num, total))
-
-
-def rebind_retracted_ars(portal):
-    """Rebind the ARs automatically generated because of the retraction of their
-    parent to the new field 'Retracted'. The field used until now
-    'ParentAnalysisRequest' will be used for partitioning
-    """
-    logger.info("Rebinding retracted/invalidated ARs")
-
-    # Walk through the Analysis Requests that were generated because of a
-    # retraction, get the source AR and rebind the fields
-    relationship = "AnalysisRequestChildAnalysisRequest"
-    ref_catalog = api.get_tool(REFERENCE_CATALOG)
-    retests = ref_catalog(relationship=relationship)
-    total = len(retests)
-    num = 0
-    for num, relation in enumerate(retests, start=1):
-        relation = relation.getObject()
-        if not relation:
-            continue
-        retest = relation.getTargetObject()
-        retracted = relation.getSourceObject()
-        retest.setRetracted(retracted)
-        # Set ParentAnalysisRequest field to None, cause we will use this field
-        # for storing Primary-Partitions relationship.
-        retest.setParentAnalysisRequest(None)
-
-        # Remove the relationship!
-        relation.aq_parent.manage_delObjects([relation.id])
-
-        if num % 100 == 0:
-            logger.info("Rebinding retracted ARs: {0}/{1}".format(num, total))
-
-
-    logger.info("{} retracted ARs have been rebinded".format(num))