+++ bika/lims/workflow/analysisrequest/events.py
@@ -5,8 +5,7 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
-from DateTime import DateTime
-
+from bika.lims.utils.analysisrequest import create_retest
 from bika.lims.workflow import doActionFor
 from bika.lims.workflow import getCurrentState
 
@@ -147,6 +146,15 @@ def after_retract(obj):
     ans = obj.getAnalyses(full_objects=True)
     for analysis in ans:
         doActionFor(analysis, 'retract')
+
+
+def after_invalidate(obj):
+    """Function triggered after 'invalidate' transition for the Analysis
+    Request passed in is performed. Creates a retest
+    :param obj: Analysis Request affected by the transition
+    :type obj: AnalysisRequest
+    """
+    create_retest(obj)
 
 
 def after_attach(obj):